+++ Sources/IKEHGitHubAPIClient/Entity/Response/SearchResponse.swift
@@ -15,4 +15,21 @@ public struct SearchResponse<Item>: Sendable, Codable, PagingResponse where Item
     // MARK: - PagingResponse
     
     public var relationLink: RelationLink? // ページング情報
+    
+    // MARK: - LifeCycle
+        
+    /// イニシャライザ
+    /// - Parameters:
+    ///   - totalCount: 項目の総数
+    ///   - items: 項目の配列
+    ///   - relationLink: ページング情報
+    public init(
+        totalCount: Int,
+        items: [Item],
+        relationLink: RelationLink? = nil
+    ) {
+        self.totalCount = totalCount
+        self.items = items
+        self.relationLink = relationLink
+    }
 }
